{{cookiecutter.project_slug}}/requirements.txt
asgiref==3.7.2
black==24.1.1
click==8.1.7
dj-database-url==2.1.0
Django==4.2.9
django-cors-headers==4.3.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.9.0
